# src/main/java/com/bank/creditservice/dto/BaseResponse.java
package com.bank.creditservice.dto;


import lombok.*;

@Data
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BaseResponse<T> {
    private int status;
    private String message;
    private T data;
}