# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Credit Service API
  description: API for managing credits and credit cards in the banking system
  version: 1.0.0
paths:
  /api/credits:
    post:
      summary: Create a new credit
      operationId: createCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '201':
          description: Credit created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all credits
      operationId: getAllCredits
      responses:
        '200':
          description: Credits retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditListResponse'
        '404':
          description: No credits found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credits/{creditId}:
    get:
      summary: Get credit by ID
      operationId: getCreditById
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a credit
      operationId: updateCredit
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credit'
      responses:
        '200':
          description: Credit updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a credit
      operationId: deleteCredit
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credits/customer/{customerId}:
    get:
      summary: Get credits by customer ID
      operationId: getCreditsByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credits for customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditListResponse'
        '404':
          description: No credits found for customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credit-cards:
    post:
      summary: Create a new credit card
      operationId: createCreditCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCard'
      responses:
        '201':
          description: Credit Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all credit cards
      operationId: getAllCreditCards
      responses:
        '200':
          description: Credit Cards retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardListResponse'
        '404':
          description: No credit cards found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credit-cards/{creditCardId}:
    get:
      summary: Get credit card by ID
      operationId: getCreditCardById
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit Card details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a credit card
      operationId: updateCreditCard
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCard'
      responses:
        '200':
          description: Credit Card updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a credit card
      operationId: deleteCreditCard
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit Card deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/credit-cards/customer/{customerId}:
    get:
      summary: Get credit cards by customer ID
      operationId: getCreditCardsByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit Cards for customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardListResponse'
        '404':
          description: No credit cards found for customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          nullable: true
    Credit:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        creditType:
          type: string
          enum: [PERSONAL, BUSINESS]
        amount:
          type: number
          format: decimal
        remainingBalance:
          type: number
          format: decimal
        interestRate:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    CreditResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Credit'
    CreditListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Credit'
    CreditCard:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        cardType:
          type: string
          enum: [PERSONAL_CREDIT_CARD, BUSINESS_CREDIT_CARD]
        creditLimit:
          type: number
          format: decimal
        availableBalance:
          type: number
          format: decimal
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    CreditCardResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/CreditCard'
    CreditCardListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/CreditCard'